requirements.txt
PyQt5
PyQtWebEngine
google-generativeai
pymavlink
numpy
flask
flask-cors
requests
python-dotenv